robot/requirements.txt
beautifulsoup4==4.13.3
PyMuPDF
python-dotenv==1.1.0
Requests==2.32.3
lxml